Solution to day 1
calcolaDistanza is the solution to the silver part of the problem. It has a time complexity of O(n  log(n)) where n is the size of the given lists. I first create a shallow copy of each list and sorting it, then I calculate the absolute value of each subtraction between the numbers and reduce it with sum. 
calcolaSomiglianza is for the solution to the gold part of the problem. By importing Counter, I made a dict where each element in the second list is counted with the element as the key. I then made that horrible one liner return where we are reducing with sum again, this time though we reduce the multiplication of how many times a given element in the first list is present in the second one. If the element is not found, we multiply it by 0 to omit it.