Solution to day 2
calcolaReattore is the solution to the silver-part of the exercise. It's an O(n) solution wher n is the size of the report list, it getsw the length of the list and uses it to start a loop. It calculates the difference between the current report and the next one (we don't care about the sign because it's an absolute value) and, if it's not between 1 and 3, returns 0 otherwise it keeps checking wether the current evaluation is increasing or decreasing. After the for loop, if neither the decreasing or increasing boolean value is still True, it returns 0 because it means that the list is unordered. Otherwise, it returns 1.
calcolaReattoreDampener is the solution to the gold-part of the exercise. It's an O(n*2) solution that calls the older solution first and, if it returns 1, it also returns 1. Otherwise, it starts slicing the report removing one element at a time and calling the calcolaReattore function. "Slow and steady wins the race", they say. 

The last part of the exercise is, as always just file reading and calling the functions to increment a value that is then displayed to give the solution.