Solution to day 4
This one probably was the hardest so far. 
We split the giant string into an array where each cell of the array is a line. We know that each string is N character long and the whole text is made of N string, so it's a NxN structure.
We declare a "dirs" array of tuples, each represents the direction in which we are analyzing the character at hand (this point will become clearer further down the road) and declare our counter.
We start to loop through each character in the list (note that the splitlines() function did the work for us on removing the "\n" characters) and then, for each of the 8 possible directions, we check if the character can lead to the word "XMAS", adding 1 to our counter.

For the 2nd part of the exercise, we again iterate through each character and, if it is equal to "A" (meaning it is a candidate for the center of a cross between 2 "MAS" strings), we check wether the 2 3-chars-long diagonals passing for this point are equal to "MAS" in any direction. Unfortunately, my solution to this is not really "pythonian".